+++ docs/.vitepress/config/theme.ts
@@ -181,6 +181,10 @@ export function getLocaleConfig(lang: string) {
               text: 'defineStyleX',
               link: `/define-stylex`,
             },
+            {
+              text: 'jsxMacros',
+              link: `/jsx-macros`,
+            },
           ],
         },
       ],