+++ packages/jsx-macros/src/core/define-model.ts
@@ -1,5 +1,4 @@
 import { importHelperFn, type MagicStringAST } from '@vue-macros/common'
-import type { OptionsResolved } from '..'
 import { useModelHelperId } from './helper'
 import type { CallExpression } from '@babel/types'
 
@@ -7,18 +6,10 @@ export function transformDefineModel(
   node: CallExpression,
   propsName: string,
   s: MagicStringAST,
-  options: OptionsResolved,
 ): void {
   s.overwriteNode(
     node.callee,
-    importHelperFn(
-      s,
-      0,
-      'useModel',
-      useModelHelperId,
-      false,
-      options.defineModel.alias[0],
-    ),
+    importHelperFn(s, 0, 'useModel', useModelHelperId),
   )
   s.appendRight(
     node.arguments[0]?.start || node.end! - 1,